requirements.txt
# HTTP and web scraping
aiohttp>=3.8.5
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.1
lxml_html_clean>=0.4.2
feedparser>=6.0.10
newspaper3k>=0.2.8

# Telegram
telethon>=1.29.0

# AI / OpenAI
openai>=1.3.7
tiktoken==0.11.0
yake>=0.4.8

# Data / Excel
pandas>=2.3.2
numpy>=1.24.4
openpyxl>=3.1.0
numexpr>=2.10.2
bottleneck>=1.5.0
python-dateutil>=2.8.2
tzdata>=2025.2

# Environment management
python-dotenv>=1.0.0
